+++ tests/test_joints.py
@@ -7,7 +7,7 @@ Created on Fri Apr 16 17:19:47 2021
 """
 
 import unittest
-from math import sqrt
+import math
 from pylinkage.exceptions import UnbuildableError
 from pylinkage.interface.joint import Pivot, Fixed
 
@@ -20,32 +20,40 @@ class TestPivot(unittest.TestCase):
     def test_buildable(self):
         """Upper intersect test."""
         pivot2 = Pivot(1, 0)
-        pivot3 = Pivot(y=1, joint0=self.pivot1, joint1=pivot2,
-                       distance0=1, distance1=1)
+        pivot3 = Pivot(
+            y=1, joint0=self.pivot1, joint1=pivot2,
+            distance0=1, distance1=1
+        )
         pivot3.reload()
-        self.assertEqual(pivot3.coord(), (.5, sqrt(.75)))
+        self.assertEqual(pivot3.coord(), (.5, math.sqrt(.75)))
 
     def test_under_intersect(self):
         """Under intersect test."""
         pivot2 = Pivot(1, 0)
-        pivot3 = Pivot(y=-1, joint0=self.pivot1, joint1=pivot2,
-                       distance0=1, distance1=1)
+        pivot3 = Pivot(
+            y=-1, joint0=self.pivot1, joint1=pivot2,
+            distance0=1, distance1=1
+        )
         pivot3.reload()
-        self.assertEqual(pivot3.coord(), (.5, -sqrt(.75)))
+        self.assertEqual(pivot3.coord(), (.5, -math.sqrt(.75)))
 
     def test_limit_intersect(self):
         """Test system almost breaking."""
         pivot2 = Pivot(2, 0)
-        pivot3 = Pivot(y=1, joint0=self.pivot1, joint1=pivot2,
-                       distance0=1, distance1=1)
+        pivot3 = Pivot(
+            y=1, joint0=self.pivot1, joint1=pivot2,
+            distance0=1, distance1=1
+        )
         pivot3.reload()
         self.assertEqual(pivot3.coord(), (1, 0))
 
     def test_no_intersect(self):
         """Test system almost breaking."""
         pivot2 = Pivot(0, 3)
-        pivot3 = Pivot(y=1, joint0=self.pivot1, joint1=pivot2,
-                       distance0=1, distance1=1)
+        pivot3 = Pivot(
+            y=1, joint0=self.pivot1, joint1=pivot2,
+            distance0=1, distance1=1
+        )
         with self.assertRaises(UnbuildableError):
             pivot3.reload()
 
@@ -58,7 +66,9 @@ class TestFixed(unittest.TestCase):
     def test_pos(self):
         """Test Fixed_Joint positioning."""
         pivot2 = Pivot(1, 0)
-        fixed = Fixed(joint0=self.pivot1, joint1=pivot2, angle=0, distance=1)
+        fixed = Fixed(
+            joint0=self.pivot1, joint1=pivot2, angle=0, distance=1
+        )
         fixed.reload()
         self.assertEqual(fixed.coord(), (1, 0))