+++ CHANGELOG.md
@@ -16,6 +16,7 @@ based on the [Strider Linkage](https://www.diywalkers.com/strider-linkage-plans.
 optimization, related to ([#5](https://github.com/HugoFara/pylinkage/issues/5)).
   - ``Agent`` is immutable and inherits from a namedtuple. It is recommended to use it, as it is a bit faster.
   - ``MutableAgent`` is mutable. It may be deprecated/removed if ``Agent`` is satisfactory.
+- Very basic tests cases for a Linkage.
 
 ### Changed